move groupingComponentListItem to groupingComponentFrames.tsx
