+++ static/app/components/events/groupingInfo/groupingComponent.tsx
@@ -48,19 +48,6 @@ const GroupingComponentList = styled('ul')<{isInline: boolean}>`
   }
 `;
 
-export const GroupingComponentListItem = styled('li')<{isCollapsible?: boolean}>`
-  padding: 0;
-  margin: ${space(0.25)} 0 ${space(0.25)} ${space(1.5)};
-
-  ${p =>
-    p.isCollapsible &&
-    css`
-      border-left: 1px solid ${p.theme.innerBorder};
-      margin: 0 0 -${space(0.25)} ${space(1)};
-      padding-left: ${space(0.5)};
-    `}
-`;
-
 export const GroupingValue = styled('code')<{
   valueType: string;
   contributes?: boolean;
